+++ rds-20140815/core/Models/RunRCInstancesRequest.cs
@@ -63,6 +63,10 @@ namespace AlibabaCloud.SDK.Rds20140815.Models
         [Validation(Required=false)]
         public bool? AutoUseCoupon { get; set; }
 
+        [NameInMap("BusinessInfo")]
+        [Validation(Required=false)]
+        public string BusinessInfo { get; set; }
+
         /// <summary>
         /// <para>The client token that is used to ensure the idempotence of the request. You can use the client to generate the token, but you must make sure that the token is unique among different requests. The token can contain only ASCII characters and cannot exceed 64 characters in length.</para>
         ///